+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 
 ### Fixed
+- Rename `CurveExportSVGPanel` to `DATA_PT_CurveExportSVG`
 
 ### Security